# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "host": "localhost:8000",
    "basePath": "/api/v1",
...
    "schemes": ["http"],
    "security": [
        {
            "Bearer": [""]
        }
    ],
    "paths": {
...
        "/createTask": {
            "post": {
                "tags": ["Task"],
                "summary": "Create a task",
                "description": "",
                "operationId": "createTask",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                },
                                "project_id": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                },
                                "user_id": {
                                    "type": "string",
                                    "default": ""
                                },
                                "description": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/assignTask/{taskId}": {
            "put": {
                "tags": ["Task"],
                "summary": "Assign Task",
                "description": "",
                "operationId": "assignTask",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskId",
                        "description": "task_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "user_id": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/updateStatus/{taskId}": {
            "put": {
                "tags": ["Task"],
                "summary": "Update task status",
                "description": "",
                "operationId": "updateStatus",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskId",
                        "description": "task_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/updateTask/{taskId}": {
            "put": {
                "tags": ["Task"],
                "summary": "Update task attribute",
                "description": "",
                "operationId": "updateTask",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskId",
                        "description": "task_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                },
                                "description": {
                                    "type": "string",
                                    "default": "",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}